requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
huggingface_hub
python-multipart
requests